update: rename SpriteAddBtn as AssetAddBtn
