+++ spx-gui/src/components/sprite-list/AssetAddBtn.vue
